pyproject.toml
[tool.poetry]
name = "django-auth0-user"
version = "0.14.0"
description = "Django Authentication and Authorisation using Auth0 and Python Social Auth"
authors = ["Samuel Bishop <sam@psyx.co>"]
packages = [
    { include = "django_auth0_user", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.6"
social-auth-core = {version = "^3.2.0", extras = ["openidconnect"]}
social-auth-app-django = "^3.1.0"
cached-property = "^1.5.1"
python-jose = {git = "https://github.com/mpdavis/python-jose.git"}
djangorestframework = {version = "^3.10.3", optional = true}
djangorestframework-jwt = {version = "^1.11.0", optional = true}
pyjwt = {version = "^1.7.1", optional = true}

[tool.poetry.dev-dependencies]
auth0-python = "^3.9.1"
django-environ = "^0.4.5"
djangorestframework = "^3.10.3"
djangorestframework-jwt = "^1.11.0"
django-extensions = "^2.2.3"
django-debug-toolbar = "^2.0"
pytest = "^5.2.0"
pytest-django = "^3.5.1"
selenium = "^3.141.0"
mimesis = "^3.3.0"
retryz = "^0.1.9"
httpretty = "^0.9.6"
pytest-server-fixtures = "^1.7.0"
pytest-selenium = "^1.17.0"
unittest2 = "^1.1.0"
django-cors-headers = "^3.1.1"

[tool.poetry.extras]
drf= ["djangorestframework", "djangorestframework-jwt", "pyjwt"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
